requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
weaviate-client==4.6.4
python-dotenv==1.0.1
pytest==8.0.2
langchain==0.3.9
langchain-openai==0.3.3
langgraph==0.3.0
langgraph-checkpoint==2.1.1



